1. file type insert file-count 2. Refactored batch service
