+++ routers/graphql-router.js
@@ -11,14 +11,9 @@ const {DatabaseConnector} = require("../crdc-datahub-database-drivers/database-c
 const {EmailService} = require("../services/email");
 const {NotifyUser} = require("../services/notify-user");
 const {User} = require("../crdc-datahub-database-drivers/services/user");
-const {Organization} = require("../crdc-datahub-database-drivers/services/organization");
 const {ApprovedStudiesService} = require("../services/approved-studies");
 const {BatchService} = require("../services/batch-service");
 const {S3Service} = require("../crdc-datahub-database-drivers/services/s3-service");
-const {verifySession} = require("../verifier/user-info-verifier");
-const {verifyBatch} = require("../verifier/batch-verifier");
-const {BATCH} = require("../crdc-datahub-database-drivers/constants/batch-constants");
-const ERROR = require("../constants/error-constants");
 const {SubmissionService} = require("../crdc-datahub-database-drivers/services/submission-service");
 
 const schema = buildSchema(require("fs").readFileSync("resources/graphql/crdc-datahub.graphql", "utf8"));
@@ -28,7 +23,6 @@ let root;
 dbConnector.connect().then(() => {
     const applicationCollection = new MongoDBCollection(dbConnector.client, DATABASE_NAME, APPLICATION_COLLECTION);
     const userCollection = new MongoDBCollection(dbConnector.client, DATABASE_NAME, USER_COLLECTION);
-    const organizationCollection = new MongoDBCollection(dbConnector.client, DATABASE_NAME, ORGANIZATION_COLLECTION);
     const emailService = new EmailService(config.email_transport, config.emails_enabled);
     const notificationsService = new NotifyUser(emailService);
     const emailParams = {url: config.emails_url, officialEmail: config.official_email, inactiveDays: config.inactive_application_days};
@@ -37,7 +31,12 @@ dbConnector.connect().then(() => {
 
     const approvedStudiesCollection = new MongoDBCollection(dbConnector.client, DATABASE_NAME, APPROVED_STUDIES_COLLECTION);
     const approvedStudiesService = new ApprovedStudiesService(approvedStudiesCollection);
-    const dataInterface = new Application(logCollection, applicationCollection, approvedStudiesService, new Organization(organizationCollection), userService, dbService, notificationsService, emailParams);
+    const submissionCollection = new MongoDBCollection(dbConnector.client, DATABASE_NAME, SUBMISSION_COLLECTION);
+    const submissionService = new SubmissionService(submissionCollection);
+    const s3Service = new S3Service();
+    const batchCollection = new MongoDBCollection(dbConnector.client, DATABASE_NAME, BATCH_COLLECTION);
+    const batchService = new BatchService(s3Service, batchCollection, config.submission_aws_bucket_name);
+    const dataInterface = new Application(logCollection, applicationCollection, approvedStudiesService, submissionService, batchService, userService, dbService, notificationsService, emailParams);
     root = {
         version: () => {return config.version},
         saveApplication: dataInterface.saveApplication.bind(dataInterface),
@@ -50,57 +49,10 @@ dbConnector.connect().then(() => {
         rejectApplication: dataInterface.rejectApplication.bind(dataInterface),
         reopenApplication: dataInterface.reopenApplication.bind(dataInterface),
         deleteApplication: dataInterface.deleteApplication.bind(dataInterface),
-        createBatch: async (params, context) => {
-            verifySession(context)
-                .verifyInitialized();
-            verifyBatch(params)
-                .isUndefined()
-                .notEmpty()
-                .type([BATCH.TYPE.METADATA, BATCH.TYPE.FILE])
-            // Optional metadata intention
-            if (params?.metadataIntention) {
-                verifyBatch(params)
-                    .metadataIntention([BATCH.INTENTION.NEW]);
-            }
-            const submissionCollection = new MongoDBCollection(dbConnector.client, DATABASE_NAME, SUBMISSION_COLLECTION);
-            const submissionService = new SubmissionService(submissionCollection);
-            await verifyBatchPermission(submissionService, dbConnector, params.submissionID, context.userInfo);
-            const s3Service = new S3Service();
-            const batchCollection = new MongoDBCollection(dbConnector.client, DATABASE_NAME, BATCH_COLLECTION);
-            const batchService = new BatchService(s3Service, batchCollection, config.submission_aws_bucket_name);
-            return await batchService.createBatch(params, context?.userInfo?.organization?.orgName);
-        }
+        createBatch: dataInterface.createBatch.bind(dataInterface),
     };
 });
 
-const verifyBatchPermission= async(submissionService, dbConnector, submissionID, userInfo) => {
-    const collectionNames = [LOG_COLLECTION, USER_COLLECTION];
-    const collections = collectionNames.map(name => new MongoDBCollection(dbConnector.client, DATABASE_NAME, name));
-    const [logCollection, userCollection] = collections;
-    const userService = new User(userCollection, logCollection);
-    // verify submission owner
-    const aSubmission = await submissionService.findByID(submissionID);
-    if (!aSubmission) {
-        throw new Error(ERROR.SUBMISSION_NOT_EXIST);
-    }
-    const aUser = await userService.getUserByID(aSubmission?.submitterID);
-    if (isPermittedUser(aUser, userInfo)) {
-        return;
-    }
-    // verify submission's organization owner by an organization name
-    const organizationOwners = await userService.getOrgOwnerByOrgName(aSubmission?.organization);
-    // organization owner only one
-    const aOrganizationOwner = organizationOwners?.length > 0 ? organizationOwners[0] : null;
-    const aOrgUser = (aOrganizationOwner) ? await userService.getUserByID(aOrganizationOwner?._id) : null;
-    if (aOrganizationOwner && isPermittedUser(aOrgUser, userInfo)) {
-        return;
-    }
-    throw new Error(ERROR.INVALID_BATCH_PERMISSION);
-}
-
-const isPermittedUser = (aTargetUser, userInfo) => {
-    return aTargetUser?.email === userInfo.email && aTargetUser?.IDP === userInfo.IDP
-}
 
 
 module.exports = (req, res) => {
