+++ services/application.js
@@ -13,17 +13,20 @@ const {CreateApplicationEvent, UpdateApplicationStateEvent} = require("../crdc-d
 const ROLES = USER_CONSTANTS.USER.ROLES;
 const config = require('../config');
 const {parseJsonString} = require("../crdc-datahub-database-drivers/utility/string-utility");
+const {verifyBatch} = require("../verifier/batch-verifier");
+const {BATCH} = require("../crdc-datahub-database-drivers/constants/batch-constants");
 
 class Application {
-    constructor(logCollection, applicationCollection, approvedStudiesService, organizationService, userService, dbService, notificationsService, emailParams) {
+    constructor(logCollection, applicationCollection, approvedStudiesService, submissionService, batchService, userService, dbService, notificationsService, emailParams) {
         this.logCollection = logCollection;
         this.applicationCollection = applicationCollection;
         this.approvedStudiesService = approvedStudiesService;
-        this.organizationService = organizationService;
         this.userService = userService;
         this.dbService = dbService;
         this.notificationService = notificationsService;
         this.emailParams = emailParams;
+        this.submissionService = submissionService;
+        this.batchService = batchService;
     }
 
     async getApplication(params, context) {
@@ -412,6 +415,47 @@ class Application {
             url: this.emailParams.url
         })
     }
+
+    async createBatch(params, context) {
+        verifySession(context)
+            .verifyInitialized();
+        verifyBatch(params)
+            .isUndefined()
+            .notEmpty()
+            .type([BATCH.TYPE.METADATA, BATCH.TYPE.FILE])
+        // Optional metadata intention
+        if (params?.metadataIntention) {
+            verifyBatch(params)
+                .metadataIntention([BATCH.INTENTION.NEW]);
+        }
+        await verifyBatchPermission(this.submissionService, this.userService, params.submissionID, context.userInfo);
+        return await this.batchService.createBatch(params, context?.userInfo?.organization?.orgName);
+    }
+}
+
+const verifyBatchPermission= async(submissionService, userService, submissionID, userInfo) => {
+    // verify submission owner
+    const aSubmission = await submissionService.findByID(submissionID);
+    if (!aSubmission) {
+        throw new Error(ERROR.SUBMISSION_NOT_EXIST);
+    }
+    const aUser = await userService.getUserByID(aSubmission?.submitterID);
+    if (isPermittedUser(aUser, userInfo)) {
+        return;
+    }
+    // verify submission's organization owner by an organization name
+    const organizationOwners = await userService.getOrgOwnerByOrgName(aSubmission?.organization);
+    // organization owner only one
+    const aOrganizationOwner = organizationOwners?.length > 0 ? organizationOwners[0] : null;
+    const aOrgUser = (aOrganizationOwner) ? await userService.getUserByID(aOrganizationOwner?._id) : null;
+    if (aOrganizationOwner && isPermittedUser(aOrgUser, userInfo)) {
+        return;
+    }
+    throw new Error(ERROR.INVALID_BATCH_PERMISSION);
+}
+
+const isPermittedUser = (aTargetUser, userInfo) => {
+    return aTargetUser?.email === userInfo.email && aTargetUser?.IDP === userInfo.IDP
 }
 
 function formatApplicantName(userInfo){
